requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
distlib==0.3.1
Django==3.0.8
django-cors-headers==3.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
filelock==3.0.12
idna==2.10
isort==5.5.2
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.6
pylint==2.6.0
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.3
virtualenv==20.0.29
wrapt==1.12.1
